Update to use Sparkle submodule rather than Cocoapods
A version of Sparkle with sandboxing support is necessary, and the Sparkle
maintainers have made clear that this cannot be expected for some time. So,
I am temporary switching to using a custom fork with this support included,
and I will switch back to Cocoapods once sandboxing support is included in
Sparkle's master.