Clear up an ambiguous option
The checkbox indicating that the ISO to install is for "Mac" has been renamed to indicate whether or not the Ubuntu-based distribution lacks an EFI-enabled kernel. This states its purpose more clearly and allows it to be seen as relevant to not only old ISOs of Ubuntu, but also modern versions of Mint which still lack a kernel whose filename ends in .efi.